Fix configuration documentation for MQTT and message brokers
The configuration documentation for MQTT and message brokers in Mainflux has been updated to provide clearer instructions on how to handle different MQTT and message brokers in the dev environment. The documentation now includes information on the two types of brokers used by Mainflux, namely `MQTT_BROKER` and `MESSAGE_BROKER`, and the options available for each. Specifically, it clarifies that `MQTT_BROKER` can be either `vernemq` or `nats`, while `MESSAGE_BROKER` can be `nats`.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>