Enhance documentation for the certs service
The documentation for the certs service had incorrect information about the different modes of operation. Specifically, the section about development mode was mistakenly labeled as PKI mode. This commit corrects the labeling and provides a more accurate description of PKI mode.

The commit also includes additional information on setting up Vault for PKI mode and provides a link to the relevant documentation.

Changes made:
- Corrected labeling of the modes of operation in the certs service documentation
- Added information on setting up Vault for PKI mode
- Provided a link to the documentation for setting up Vault as a Certificate Authority (CA)

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>